Organize and naming components into sub folder by area of application
