+++ webapp/src/components/CreateUser/CreateUserPage.vue
@@ -1,0 +1,6 @@
+<template>
+
+</template>
+<script>
+</script>
+<style scoped></style>
