+++ webapp/src/components/Home/HomeContent.vue
@@ -3,31 +3,31 @@
         <ul>
            <li>
                 <div id="contact-friend">
-                    <button><img src="../assets/img/user-blue-ico.png" alt="friend"/></button>
+                    <button><img src="/src/assets/img/user-blue-ico.png" alt="friend"/></button>
                     <h7>Friend 1</h7>
                 </div>
            </li> 
            <li>
                 <div id="contact-group">
-                    <button><img src="../assets/img/team-ico.png" alt="group 1"/></button>
+                    <button><img src="/src/assets/img/team-ico.png" alt="group 1"/></button>
                     <h7>Group 1</h7>
                 </div>
            </li>
            <li>
                 <div id="contact-group">
-                    <button><img src="../assets/img/groups-ico.png" alt="group 2"/></button>
+                    <button><img src="/src/assets/img/groups-ico.png" alt="group 2"/></button>
                     <h7>Group 2</h7>
                 </div>
            </li>
            <li>
                 <div id="contact-group">
-                    <button><img src="../assets/img/team-ico.png" alt="group 3"/></button>
+                    <button><img src="/src/assets/img/team-ico.png" alt="group 3"/></button>
                     <h7>Group</h7>
                 </div>
             </li>
            <li>
                 <div id="contact-friend">
-                    <button><img src="../assets/img/user-blue-ico.png" alt="friend 2"/></button>
+                    <button><img src="/src/assets/img/user-blue-ico.png" alt="friend 2"/></button>
                     <h7>Friend 2</h7>
                 </div>    
             </li>
