+++ webapp/src/components/Home/HomePage.vue
@@ -1,0 +1,16 @@
+<script setup>
+    import MyNav from '../NavBar.vue';
+    import HomeContent from './HomeContent.vue';
+    import SCLogo from '../SCLogo.vue';
+</script>
+
+<template>
+    <div id="wrapper-home">
+        <MyNav></MyNav>
+        <HomeContent></HomeContent>
+        <SCLogo></SCLogo>
+    </div>
+</template>
+
+<style scoped>
+</style>
