--- webapp/src/components/MyHome.vue
@@ -1,16 +1,0 @@
-<script setup>
-    import MyNav from './MyNav.vue';
-    import HomeMain from './HomeMain.vue';
-    import MyLogo from './MyLogo.vue';
-</script>
-
-<template>
-    <div id="wrapper-home">
-        <MyNav></MyNav>
-        <HomeMain></HomeMain>
-        <MyLogo></MyLogo>
-    </div>
-</template>
-
-<style>
-</style>
