+++ webapp/src/components/NavBar.vue
