+++ webapp/src/components/SCLogo.vue
